Extract remaining 5 clusters from process_control.py
Third and final major extraction from modules/process_control.py.
Rests on the process_control_state foundation from PR #1523.

## What moved

Five thematic modules created, each self-contained on top of the
bottom-layer process_control_state:

### modules/process_maintenance.py (~208 lines)
9 functions covering maintenance-window resolution:
- parse_maintenance_window_minutes, is_within_maintenance_window
- get_maintenance_window_from_db, get_plex_credentials_from_db
- get_maintenance_window_live, get_active_maintenance_lookup_config_name
- resolve_maintenance_window_live, resolve_maintenance_window_from_db
- refresh_maintenance_window_availability

### modules/process_pending_start.py (~70 lines)
4 functions guarding the pending-Kometa-start queue:
- set/peek/pop/clear_pending_kometa_start

### modules/process_discovery.py (~111 lines)
4 psutil scan functions:
- find_running_kometa_process(es)
- find_running_imagemaid_process(es)

### modules/process_lifecycle.py (~244 lines)
6 imperative subprocess verbs:
- stop_process_tree, suspend_process_tree, resume_process_tree
- launch_kometa_command, launch_imagemaid_command
- reset_imagemaid_runtime_env

### modules/process_run_context.py (~168 lines)
8 run-context accessors (Kometa + ImageMaid):
- extract_selected_libraries
- update/get/clear_run_context
- normalize_imagemaid_command_text
- update/get/clear_imagemaid_run_context

## modules/process_control.py: 839 -> 349 (-490 / -58.4%)

Cleaned up newly-unused imports (re, shlex, subprocess, sys,
Path, database, persistence, has_request_context, session) whose
users all moved out.  The 349-line file now contains only:

- Module docstring explaining the split
- 8 re-export blocks (metrics + state + markers + maintenance +
  pending_start + discovery + lifecycle + run_context)
- The single remaining function: maintenance_guard_loop (~200 lines).

maintenance_guard_loop stays put because it is the top-level
orchestrator: it ties every other cluster together (calls
maintenance-window resolvers, reads/writes MAINTENANCE_STATE,
launches pending starts, suspends/resumes processes, writes
maintenance markers).  Moving it would require solving a
6-way circular import for one function -- not worth it.

## Sprint 4 cumulative process_control.py totals

- Sprint start:           1259 lines
- After PR #1522:         1155 lines (-104, metrics extraction)
- After PR #1523:          839 lines (-316, state + markers)
- After this PR:           349 lines (-490, everything else)
- Total reduction:        -910 lines / -72.3%

## Backward compatibility

process_control.py re-exports every moved name via 8 re-export
blocks.  All 43 imports in quickstart.py (using X as _X aliases)
continue to work unchanged.  No test-side changes needed --
existing patches on qs_module._X aliases route through
quickstart.__dict__, which doesn't care where the underlying
function lives.

## Verification

AST-verified: all 31 moved functions in this PR are byte-for-byte
identical to origin/develop.  Full test suite passes (0 failures).
Ruff clean.  Black clean.
